Prevent save of clashing users when username is case-insensitive #7
For user type models with a case-insensitive username surrogate field,
as indicated by the `IS_USERNAME_CASE_INSENSITIVE` class attribute,
perform a check before saving new users to ensure that the username
provided doesn't clash with an existing username differing only by
case.

In particular, this change will prevent creation of `EmailUser`
accounts with slight case variations of the email address.

PENDING: Do the same checks in the admin forms to return a nicer
error message to the user instead of a 500 error. This should be
a last-ditch effort to maintain data integrity.